# build/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "mongoose.Types.ObjectId": {
        "type": "string"
      },
...
      "ProductState": {
        "description": "************* 产品相关 ***************",
        "enum": ["on-line", "off-line"],
        "type": "string"
      },
      "Record_string.any_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
      "LanguageItemType": {
        "properties": {
          "languageMap": {
            "$ref": "#/components/schemas/Record_string.any_"
          },
          "language": {
            "type": "string"
          }
        },
        "required": ["languageMap", "language"],
        "type": "object"
      },
      "ProductDTO": {
        "properties": {
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录创建时间"
          },
          "modifyTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录修改时间"
          },
          "deleteTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "记录删除时间"
          },
          "languageList": {
            "items": {
              "$ref": "#/components/schemas/LanguageItemType"
            },
            "type": "array"
          },
          "cates": {
            "items": {
              "$ref": "#/components/schemas/mongoose.Types.ObjectId"
            },
            "type": "array"
          },
          "productName": {
            "type": "string"
          },
          "masterPicture": {
            "type": "string"
          },
          "descPic": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "slug": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ProductState"
          },
          "richText": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "allOf": [
              {
                "$ref": "#/components/schemas/mongoose.Types.ObjectId"
              }
            ],
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "activityPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "sales": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "score": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "required": ["languageList", "cates", "productName", "masterPicture", "descPic", "slug", "state"],
        "type": "object",
        "additionalProperties": false
      },
      "PageListType_ProductDTO_": {
        "properties": {
          "pages": {
            "type": "number",
            "format": "double"
          },
          "pageIndex": {
            "type": "number",
            "format": "double"
          },
          "pageSize": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "list": {
            "items": {
              "$ref": "#/components/schemas/ProductDTO"
            },
            "type": "array"
          }
        },
        "required": ["pages", "pageIndex", "pageSize", "total", "list"],
        "type": "object"
      },
      "BasePageListEntity_ProductDTO_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/PageListType_ProductDTO_"
          }
        },
        "required": ["data"],
        "type": "object",
        "additionalProperties": false
      },
...
      "BaseObjectEntity_ProductDTO_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/ProductDTO"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "BaseObjectEntity_string_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "BaseObjectEntity_unknown_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {}
        },
        "type": "object",
        "additionalProperties": false
      },
      "Boolean": {
        "properties": {},
        "type": "object",
        "additionalProperties": false
      },
      "BaseObjectEntity_Boolean_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/Boolean"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Pick_ProductDTO.slug_": {
        "properties": {
          "slug": {
            "type": "string"
          }
        },
        "required": ["slug"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "CheckSlugParams": {
        "$ref": "#/components/schemas/Pick_ProductDTO.slug_"
      },
      "Partial_ProductDTO_": {
        "properties": {
          "cates": {
            "items": {
              "$ref": "#/components/schemas/mongoose.Types.ObjectId"
            },
            "type": "array"
          },
          "productName": {
            "type": "string"
          },
          "masterPicture": {
            "type": "string"
          },
          "descPic": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "slug": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ProductState"
          },
          "richText": {
            "type": "string"
          },
          "brand": {
            "$ref": "#/components/schemas/mongoose.Types.ObjectId"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "activityPrice": {
            "type": "number",
            "format": "double"
          },
          "sales": {
            "type": "number",
            "format": "double"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "languageList": {
            "items": {
              "$ref": "#/components/schemas/LanguageItemType"
            },
            "type": "array"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录创建时间"
          },
          "modifyTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录修改时间"
          },
          "deleteTime": {
            "type": "string",
            "format": "date-time",
            "description": "记录删除时间"
          }
        },
        "type": "object",
        "description": "Make all properties in T optional"
      },
      "EditProductParams": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Partial_ProductDTO_"
          },
          {
            "properties": {
              "brandId": {
                "type": "string"
              }
            },
            "required": ["brandId"],
            "type": "object"
          }
        ]
      },
      "BaseObjectEntity_ProductDTO-or-null_": {
        "properties": {
          "status": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "message": {
            "type": "string",
            "default": "操作成功"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductDTO"
              }
            ],
            "nullable": true
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "KBMall",
    "version": "1.0.0",
    "description": "KB商城的买家端、卖家端的接口服务",
    "license": {
      "name": "ISC"
    },
    "contact": {
      "email": "kbmjj123@gmail.com",
      "name": "郑耿林"
    }
  },
  "paths": {
...
    "/product/list": {
      "get": {
        "operationId": "GetProductList",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasePageListEntity_ProductDTO_"
                }
              }
            }
          }
        },
        "description": "获取商品列表",
        "tags": ["产品模块"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "pageIndex",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/product/{id}": {
      "get": {
        "operationId": "GetAProduct",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_ProductDTO_"
                }
              }
            }
          }
        },
        "description": "获取一个商品明细信息",
        "tags": ["产品模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "operationId": "RemoveAProduct",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_string_"
                }
              }
            }
          }
        },
        "description": "删除一个商品",
        "tags": ["产品模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "EditProduct",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_ProductDTO-or-null_"
                }
              }
            }
          }
        },
        "description": "编辑商品",
        "tags": ["产品模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
    "/product/{id}/upOrDownShelves": {
      "post": {
        "operationId": "UpOrDownAProduct",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_unknown_"
                }
              }
            }
          }
        },
        "description": "上/下架一款商品",
        "tags": ["产品模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "state": {
                    "$ref": "#/components/schemas/ProductState"
                  }
                },
                "required": ["state"],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/product/slug/check": {
      "post": {
        "operationId": "CheckSlugUnique",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_Boolean_"
                }
              }
            }
          }
        },
        "description": "检查slug的唯一性",
        "tags": ["产品模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckSlugParams"
              }
            }
          }
        }
      }
    },
    "/product/publish": {
      "put": {
        "operationId": "CreateProduct",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_ProductDTO_"
                }
              }
            }
          }
        },
        "description": "发布一个商品",
        "tags": ["产品模块"],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditProductParams"
              }
            }
          }
        }
      }
    },
    "/product/{id}/quickEditPrice": {
      "post": {
        "operationId": "QuickEditPrice",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_ProductDTO-or-null_"
                }
              }
            }
          }
        },
        "description": "快速编辑商品价格",
        "tags": ["产品模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
...
    "/api/product/list": {
      "get": {
        "operationId": "GetProductList",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasePageListEntity_ProductDTO_"
                }
              }
            }
          }
        },
        "description": "获取商品列表",
        "tags": ["商城/商品模块"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "pageIndex",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/product/{id}": {
      "get": {
        "operationId": "GetProductDetail",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseObjectEntity_unknown_"
                }
              }
            }
          }
        },
        "description": "获取商品详情信息",
        "tags": ["商城/商品模块"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}